move beacon committee selection to validator top-level module
