+++ beacon-chain/rpc/eth/validator/structs.go
@@ -2,11 +2,6 @@ package validator
 
 import (
 	"encoding/json"
-	"strconv"
-
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/pkg/errors"
-	"github.com/prysmaticlabs/prysm/v4/consensus-types/primitives"
 
 	"github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/shared"
 )
@@ -94,53 +89,4 @@ type GetLivenessResponse struct {
 type Liveness struct {
 	Index  string `json:"index"`
 	IsLive bool   `json:"is_live"`
-}
-
-type BeaconCommitteeSelection struct {
-	SelectionProof []byte
-	Slot           primitives.Slot
-	ValidatorIndex primitives.ValidatorIndex
-}
-
-type beaconCommitteeSelectionJson struct {
-	SelectionProof string `json:"selection_proof"`
-	Slot           string `json:"slot"`
-	ValidatorIndex string `json:"validator_index"`
-}
-
-func (b BeaconCommitteeSelection) MarshalJSON() ([]byte, error) {
-	return json.Marshal(beaconCommitteeSelectionJson{
-		SelectionProof: hexutil.Encode(b.SelectionProof),
-		Slot:           strconv.FormatUint(uint64(b.Slot), 10),
-		ValidatorIndex: strconv.FormatUint(uint64(b.ValidatorIndex), 10),
-	})
-}
-
-func (b *BeaconCommitteeSelection) UnmarshalJSON(input []byte) error {
-	var bjson beaconCommitteeSelectionJson
-	err := json.Unmarshal(input, &bjson)
-	if err != nil {
-		return errors.Wrap(err, "failed to unmarshal beacon committee selection")
-	}
-
-	slot, err := strconv.ParseUint(bjson.Slot, 10, 64)
-	if err != nil {
-		return errors.Wrap(err, "failed to parse slot")
-	}
-
-	vIdx, err := strconv.ParseUint(bjson.ValidatorIndex, 10, 64)
-	if err != nil {
-		return errors.Wrap(err, "failed to parse validator index")
-	}
-
-	selectionProof, err := hexutil.Decode(bjson.SelectionProof)
-	if err != nil {
-		return errors.Wrap(err, "failed to parse selection proof")
-	}
-
-	b.Slot = primitives.Slot(slot)
-	b.SelectionProof = selectionProof
-	b.ValidatorIndex = primitives.ValidatorIndex(vIdx)
-
-	return nil
 }
