+++ testing/validator-mock/validator_client_mock.go
@@ -8,8 +8,9 @@ import (
 	context "context"
 	reflect "reflect"
 
+	"github.com/prysmaticlabs/prysm/v4/validator/client/iface"
+
 	gomock "github.com/golang/mock/gomock"
-	"github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/validator"
 	primitives "github.com/prysmaticlabs/prysm/v4/consensus-types/primitives"
 	eth "github.com/prysmaticlabs/prysm/v4/proto/prysm/v1alpha1"
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
@@ -83,10 +84,10 @@ func (mr *MockValidatorClientMockRecorder) EventStreamIsRunning() *gomock.Call {
 }
 
 // GetAggregatedSelections mocks base method.
-func (m *MockValidatorClient) GetAggregatedSelections(arg0 context.Context, arg1 []validator.BeaconCommitteeSelection) ([]validator.BeaconCommitteeSelection, error) {
+func (m *MockValidatorClient) GetAggregatedSelections(arg0 context.Context, arg1 []iface.BeaconCommitteeSelection) ([]iface.BeaconCommitteeSelection, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetAggregatedSelections", arg0, arg1)
-	ret0, _ := ret[0].([]validator.BeaconCommitteeSelection)
+	ret0, _ := ret[0].([]iface.BeaconCommitteeSelection)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
