+++ validator/client/beacon-api/beacon_api_validator_client.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"time"
 
-	"github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/validator"
-
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/golang/protobuf/ptypes/empty"
 	"github.com/pkg/errors"
@@ -178,6 +176,6 @@ func (c *beaconApiValidatorClient) EventStreamIsRunning() bool {
 	return c.eventHandler.running
 }
 
-func (c *beaconApiValidatorClient) GetAggregatedSelections(ctx context.Context, selections []validator.BeaconCommitteeSelection) ([]validator.BeaconCommitteeSelection, error) {
+func (c *beaconApiValidatorClient) GetAggregatedSelections(ctx context.Context, selections []iface.BeaconCommitteeSelection) ([]iface.BeaconCommitteeSelection, error) {
 	return c.getAggregatedSelection(ctx, selections)
 }
