+++ validator/client/beacon-api/beacon_committee_selections.go
@@ -5,16 +5,16 @@ import (
 	"context"
 	"encoding/json"
 
-	"github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/validator"
+	"github.com/prysmaticlabs/prysm/v4/validator/client/iface"
 
 	"github.com/pkg/errors"
 )
 
 type aggregatedSelectionResponse struct {
-	Data []validator.BeaconCommitteeSelection `json:"data"`
+	Data []iface.BeaconCommitteeSelection `json:"data"`
 }
 
-func (c *beaconApiValidatorClient) getAggregatedSelection(ctx context.Context, selections []validator.BeaconCommitteeSelection) ([]validator.BeaconCommitteeSelection, error) {
+func (c *beaconApiValidatorClient) getAggregatedSelection(ctx context.Context, selections []iface.BeaconCommitteeSelection) ([]iface.BeaconCommitteeSelection, error) {
 	body, err := json.Marshal(selections)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to marshal selections")
