+++ validator/client/iface/validator_client.go
@@ -2,13 +2,65 @@ package iface
 
 import (
 	"context"
+	"encoding/json"
+	"strconv"
 
-	"github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/validator"
+	"github.com/ethereum/go-ethereum/common/hexutil"
+	"github.com/pkg/errors"
 
 	"github.com/golang/protobuf/ptypes/empty"
 	"github.com/prysmaticlabs/prysm/v4/consensus-types/primitives"
 	ethpb "github.com/prysmaticlabs/prysm/v4/proto/prysm/v1alpha1"
 )
+
+type BeaconCommitteeSelection struct {
+	SelectionProof []byte
+	Slot           primitives.Slot
+	ValidatorIndex primitives.ValidatorIndex
+}
+
+type beaconCommitteeSelectionJson struct {
+	SelectionProof string `json:"selection_proof"`
+	Slot           string `json:"slot"`
+	ValidatorIndex string `json:"validator_index"`
+}
+
+func (b BeaconCommitteeSelection) MarshalJSON() ([]byte, error) {
+	return json.Marshal(beaconCommitteeSelectionJson{
+		SelectionProof: hexutil.Encode(b.SelectionProof),
+		Slot:           strconv.FormatUint(uint64(b.Slot), 10),
+		ValidatorIndex: strconv.FormatUint(uint64(b.ValidatorIndex), 10),
+	})
+}
+
+func (b *BeaconCommitteeSelection) UnmarshalJSON(input []byte) error {
+	var bjson beaconCommitteeSelectionJson
+	err := json.Unmarshal(input, &bjson)
+	if err != nil {
+		return errors.Wrap(err, "failed to unmarshal beacon committee selection")
+	}
+
+	slot, err := strconv.ParseUint(bjson.Slot, 10, 64)
+	if err != nil {
+		return errors.Wrap(err, "failed to parse slot")
+	}
+
+	vIdx, err := strconv.ParseUint(bjson.ValidatorIndex, 10, 64)
+	if err != nil {
+		return errors.Wrap(err, "failed to parse validator index")
+	}
+
+	selectionProof, err := hexutil.Decode(bjson.SelectionProof)
+	if err != nil {
+		return errors.Wrap(err, "failed to parse selection proof")
+	}
+
+	b.Slot = primitives.Slot(slot)
+	b.SelectionProof = selectionProof
+	b.ValidatorIndex = primitives.ValidatorIndex(vIdx)
+
+	return nil
+}
 
 type ValidatorClient interface {
 	GetDuties(ctx context.Context, in *ethpb.DutiesRequest) (*ethpb.DutiesResponse, error)
@@ -38,5 +90,5 @@ type ValidatorClient interface {
 	SubmitValidatorRegistrations(ctx context.Context, in *ethpb.SignedValidatorRegistrationsV1) (*empty.Empty, error)
 	StartEventStream(ctx context.Context) error
 	EventStreamIsRunning() bool
-	GetAggregatedSelections(ctx context.Context, selections []validator.BeaconCommitteeSelection) ([]validator.BeaconCommitteeSelection, error)
+	GetAggregatedSelections(ctx context.Context, selections []BeaconCommitteeSelection) ([]BeaconCommitteeSelection, error)
 }
